services/auth/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
orjson>=3.11.1,<3.12

SQLAlchemy[asyncio]==2.0.32
asyncpg==0.29.0

redis==5.0.4

python-dotenv==1.0.1
pydantic-settings==2.5.2

PyJWT==2.9.0
bcrypt==4.3.0

typer==0.12.3

httpx==0.27.0

opentelemetry-api>=1.24,<2
opentelemetry-sdk>=1.24,<2
opentelemetry-exporter-otlp-proto-http>=1.24,<2
opentelemetry-instrumentation-fastapi>=0.45b0,<1
opentelemetry-instrumentation-httpx>=0.45b0,<1
